Breaking filter into Filter and Algorithm classes.
Filter DOES NOT WORK.
